pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-demo"
version = "0.2.0"
description = "Hook-driven secure agent runtime — substrate for enterprise AI automation."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "dior-ai", email = "fullstack.engineer418@gmail.com" },
]
keywords = ["agents", "sandbox", "hooks", "mcp", "audit", "policy"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "rich>=13.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
# Either provider works; install whichever LLM you have a key for.
llm = [
    "anthropic>=0.45.0",
    "openai>=1.40.0",
]
anthropic = ["anthropic>=0.45.0"]
openai = ["openai>=1.40.0"]
# Operator console — Textual TUI for browsing runs and audit events.
ui = [
    "textual>=0.80.0",
]
# Real browser engine for the browser-research demo. Pulls Chromium
# (~250MB) on first run via `playwright install chromium`. The
# default browser-research demo uses the in-process FakeBrowser and
# does NOT need this extra.
browser = [
    "playwright>=1.40.0",
]

[project.scripts]
claude-demo = "claude_demo.cli.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["claude_demo*"]
